requirements.txt
torch
torchvision
transformers
Pillow
ultralytics
opencv-python
